requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.1
sqlalchemy>=2.0.0
pytest>=7.0.0
pytest-mock>=3.0.0
jieba>=0.42.1
